Add check semantics
- Adds check for hsValues
- Moves validation to a new file
